Added Return to Safeground for Damagables
This is if we didn't move the character they would take repeated damage or make the encounter a trifle. It returns them to the last safe ground they were on.

There is also a piece of code that looks for triggers that are marked unsafe and if in one it will not record  the location (good for moving spikes or spinning blades)